+++ week04/OnlineOrdering/Product.cs
@@ -2,9 +2,9 @@ public class Product
 {
     private string _name;
     private int _id;
-    private int _price;
+    private double _price;
     private int _quantity;
-    public Product(string name, int id, int price, int quantity)
+    public Product(string name, int id, double price, int quantity)
     {
         _name = name;
         _id = id;
@@ -23,7 +23,7 @@ public class Product
     }
 
 
-    public int TotalCost()
+    public double TotalCost()
     {
         return _price * _quantity;
     }
